Add details specific to Debian-based installs
Debian and derivative distributions do not configure the environment for compiling native code by default, and require the `build-essential` package to be installed first. Not having this installed can cause various strange failure-to-start-server issues which log with all manner of error messages ranging from permission errors to file-not-found ones.

This small PR updates the main readme to include a note regarding this pre-req.